Add lib to compiler to export compile_string
