+++ vm/ignite/src/repl.rs
@@ -1,5 +1,6 @@
 use anyhow::Result;
+use compiler::compile_string;
 
 pub fn ignite_repl() -> Result<()>  {
-
+    Ok(())
 }
